+++ lib/ts/recipe/emailpassword/componentOverrideContext.tsx
@@ -1,5 +1,5 @@
-import { createGenericComponentsOverrideContext } from "../../../../components/componentOverride/genericComponentOverrideContext";
-import { ComponentOverrideMap } from "../../types";
+import { createGenericComponentsOverrideContext } from "../../components/componentOverride/genericComponentOverrideContext";
+import { ComponentOverrideMap } from "./types";
 
 const [useOverrideContext, Provider, Consumer] = createGenericComponentsOverrideContext<ComponentOverrideMap>();
