+++ lib/ts/recipe/passwordless/components/features/linkClickedScreen/index.tsx
@@ -34,7 +34,7 @@ import { getLoginAttemptInfo } from "../../../utils";
 import STGeneralError from "supertokens-web-js/utils/error";
 import Session from "../../../../session/recipe";
 import SuperTokens from "../../../../../superTokens";
-import { useOverrideContext } from "../../componentOverride/componentOverrideContext";
+import { useOverrideContext } from "../../../componentOverrideContext";
 
 type PropType = FeatureBaseProps & { recipe: Recipe };
